python_bot/requirements.txt
discord.py==2.3.2
pymongo==4.6.1
flask==2.3.3
python-dotenv==1.0.0
pyyaml==6.0.1
aiohttp==3.9.1
colorlog==6.7.0
motor==3.3.2
pydantic==2.5.2
asyncio==3.4.3
datetime
wavelink==2.6.7
youtube-dl==2021.12.17
PyNaCl==1.5.0 